Many minor changes to better support modular renderers.
This uses a tr_config.h based setup so you can easily reuse any renderer
files without worrying about whether the original files point to the old
or new *.h files.

Also removed all of the unchanged files from the OA renderer so it is
easier to view and maintain.